requirements.txt
# Bluesky API
atproto>=0.0.40

# Google AI (Gemini)
google-generativeai>=0.8.0

# Vector Database
chromadb>=0.5.0

# Environment management
python-dotenv>=1.0.0

# Utilities
requests>=2.28.0
loguru>=0.7.0
rich>=13.0.0
websockets>=13.0.0
streamlit>=1.37.0
tqdm>=4.64.0

# API Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# (Optional) Remove heavy deps if not used
# pandas
# numpy
# nltk
# langchain-text-splitters
# streamlit
# pydantic
